小木棍/requirements.txt
pandas>=1.5.0
pyyaml>=6.0
requests>=2.28.0
colorama>=0.4.0
tqdm>=4.65.0
typing-extensions>=4.5.0